#34 refactor : Entity To DTO 변환 메서드 위치 올바르게 수정
