+++ src/main/java/com/idea5/four_cut_photos_map/domain/favorite/dto/response/FavoriteResponseDto.java
@@ -1,5 +1,6 @@
 package com.idea5.four_cut_photos_map.domain.favorite.dto.response;
 
+import com.idea5.four_cut_photos_map.domain.favorite.entity.Favorite;
 import com.idea5.four_cut_photos_map.domain.shop.dto.response.ResponseFavoriteShop;
 import com.idea5.four_cut_photos_map.domain.shop.entity.Shop;
 import lombok.AllArgsConstructor;
@@ -12,16 +13,16 @@ import lombok.Setter;
 @AllArgsConstructor
 @Builder
 public class FavoriteResponseDto {
-        private Long id;
-        private ResponseFavoriteShop shop;
+    private Long id;
+    private ResponseFavoriteShop shop;
 
-        public static ResponseFavoriteShop from(Shop shop) {
-                return ResponseFavoriteShop.builder()
-                        .id(shop.getId())
-                        .brand(shop.getBrand())
-                        .placeName(shop.getPlaceName())
-                        .roadAddressName(shop.getRoadAddressName())
-                        .favoriteCnt(shop.getFavoriteCnt())
-                        .build();
-        }
+    public static FavoriteResponseDto from(Favorite favorite) {
+        ResponseFavoriteShop shopDto = ResponseFavoriteShop.from(favorite.getShop());
+
+        return FavoriteResponseDto.builder()
+                .id(favorite.getId())
+                .shop(shopDto)
+                .build();
+    }
+
 }
