+++ src/main/java/com/idea5/four_cut_photos_map/domain/shop/dto/response/ResponseFavoriteShop.java
@@ -1,5 +1,6 @@
 package com.idea5.four_cut_photos_map.domain.shop.dto.response;
 
+import com.idea5.four_cut_photos_map.domain.shop.entity.Shop;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -15,4 +16,14 @@ public class ResponseFavoriteShop {
     private String placeName;
     private String roadAddressName;
     private Integer favoriteCnt;
+
+    public static ResponseFavoriteShop from(Shop shop) {
+        return ResponseFavoriteShop.builder()
+                .id(shop.getId())
+                .brand(shop.getBrand())
+                .placeName(shop.getPlaceName())
+                .roadAddressName(shop.getRoadAddressName())
+                .favoriteCnt(shop.getFavoriteCnt())
+                .build();
+    }
 }
